+++ service/src/main/java/org/grnet/cat/services/UserService.java
@@ -11,7 +11,6 @@ import org.grnet.cat.dtos.pagination.PageResource;
 import org.grnet.cat.entities.User;
 import org.grnet.cat.entities.Validation;
 import org.grnet.cat.enums.Source;
-import org.grnet.cat.enums.UserType;
 import org.grnet.cat.enums.ValidationStatus;
 import org.grnet.cat.exceptions.ConflictException;
 import org.grnet.cat.mappers.UserMapper;